an - Added filter by tag and modified seed to include tag

+++ db/migrate/20200724180244_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -7,9 +7,8 @@ end
 AddMissingUniqueIndices.class_eval do
   def self.up
     add_index ActsAsTaggableOn.tags_table, :name, unique: true
-    if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
-      remove_index ActsAsTaggableOn.taggings_table, :tag_id
-    end
+
+    remove_index ActsAsTaggableOn.taggings_table, :tag_id if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
     remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
     add_index ActsAsTaggableOn.taggings_table,
               %i[tag_id taggable_id taggable_type context tagger_id tagger_type],
@@ -18,12 +17,10 @@ AddMissingUniqueIndices.class_eval do
 
   def self.down
     remove_index ActsAsTaggableOn.tags_table, :name
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_idx'
-    add_index ActsAsTaggableOn.taggings_table, :tag_id
-    return if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
 
-    add_index ActsAsTaggableOn.taggings_table,
-              %i[taggable_id taggable_type context],
-              name: 'taggings_taggable_context_idx'
+    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_idx'
+
+    add_index ActsAsTaggableOn.taggings_table, :tag_id unless index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
+    add_index ActsAsTaggableOn.taggings_table, %i[taggable_id taggable_type context], name: 'taggings_taggable_context_idx'
   end
 end
